% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEVD.R
\docType{package}
\name{DEEVD-package}
\alias{DEEVD}
\alias{DEEVD-package}
\title{DEEVD}
\description{
Two extreme value distributions; Weibull and Gumbel kernel related functions are presented. \code{\link{Weibull}} and \code{\link{Gumbel}} present estimated values and \code{\link{plot.Weibull}} and \code{\link{plot.Gumbel}} plot the densities. While mean squared error can be calculated by using \code{\link{mse}}. Further, some related data sets are also presented.
}
\details{
Density Estimation by Extreme Value Distributions
}
\references{
\itemize{
\item Salha, R. B., El Shekh Ahmed, H. I., & Alhoubi, I. M. 2014. Hazard Rate Function Estimation Using Weibull Kernel. \emph{Open Journal of Statistics} \strong{4} (08), 650-661.
\item Khan, J. A., & Akbar, A. 2021. Density Estimation Using Gumbel Kernel Estimator. \emph{Open Journal of Statistics} \strong{11} (2), 319-328.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=DEEVD}
}

}
\author{
Javaria Ahmad Khan, Atif Akbar.
}
