% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEVD.R
\name{figGum}
\alias{figGum}
\title{Plot Density by Gumbel kernel.}
\usage{
figGum(y, k, h)
}
\arguments{
\item{y}{a numeric vector of positive values.}

\item{k}{gird points.}

\item{h}{the bandwidth}
}
\value{
Plot the density by using Gumbel Kernel
}
\description{
Plot Kernel density by using Gumbel Kernel.
}
\examples{
y<-rexp(23,1)
h<-0.79 * IQR(y) * length(y) ^ (-1/5)
figGum(y,80,h)
}
