% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEVD.R
\name{figWeibull}
\alias{figWeibull}
\title{Plot Density by Weibull kernel.}
\usage{
figWeibull(y, k, h)
}
\arguments{
\item{y}{a numeric vector of positive values.}

\item{k}{gird points.}

\item{h}{the bandwidth}
}
\value{
Plot the density by using Weibull Kernel
}
\description{
Plot Kernel density by using weibull Kernel.
}
\examples{
y<-rexp(23,1)
h<-0.79 * IQR(y) * length(y) ^ (-1/5)
figWeibull(y,80,h)
}
