% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEVD.R
\docType{data}
\name{Mississippi}
\alias{Mississippi}
\title{Flood discharge in per second from Mississippi river}
\format{
A vector with 23 observations
}
\usage{
Mississippi
}
\description{
A dataset containing the flood discharge in per second in cubic meter from Mississippi  river.
}
\references{
Gumbel, E. J. 1941. The return period of flood flows. \emph{The Annals of Mathematical Statistics}. \strong{12}, 163-190.
}
\keyword{datasets}
