% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEVD.R
\name{msegum}
\alias{msegum}
\title{Calculate Mean Square Error( MSE) when Gumbel kernel is used.}
\usage{
msegum(y, k, h, type)
}
\arguments{
\item{y}{a numeric vector of positive values.}

\item{k}{gird points.}

\item{h}{the bandwidth}

\item{type}{mention distribution of vector.If Gumbeldistribution is used scale=1 then use "Gumbel".
if use Weibull distribution with scale = 1 then use "Weibull".
if use Frechet distribution with scale=1 and shape=1 then use "Frechet".}
}
\value{
MSE
}
\description{
Calculate MSE by using Gumbel Kernel.
}
\examples{
y<-rweibull(350,1)
h<-0.79 * IQR(y) * length(y) ^ (-1/5)
msegum(y,200,h,"Weibull")
}
