% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEVD.R
\docType{data}
\name{Suicide}
\alias{Suicide}
\title{Data of control patients of Suicide study}
\format{
A vector with 86 observations
}
\usage{
Suicide
}
\description{
A dataset which contains a length of treatment spells (in days) of control patients in suicide study.
}
\references{
Silverman, B. W. 1986. \emph{Density Estimation}. Chapman & Hall/ CRC, London.
}
\keyword{datasets}
