% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEEVD.R
\docType{package}
\name{DEEVD-package}
\alias{DEEVD}
\alias{DEEVD-package}
\title{DEEVD}
\description{
Weibull and Gumbel kernel related functions are presented. \code{\link{Weibull}} and \code{\link{Gumbel}} present estimated values and \code{\link{plot.Weibull}} and \code{\link{plot.Gumbel}} plot the densities. While \code{\link{mseweibull}} and \code{\link{msegumbel}} calculate the Mean Squared Error (MSE).
}
\details{
Density Estimation by Extreme Value Distributions
}
\references{
\itemize{
\item Salha, R. B., El Shekh Ahmed, H. I., & Alhoubi, I. M. 2014. Hazard Rate Function Estimation Using Weibull Kernel. \emph{Open Journal of Statistics} \strong{4} (08), 650-661.
\item Khan, J. A.; Akbar, A. Density Estimation by Gumbel Kernel. 2020. \emph{Working paper,  Department of Statistics, Bahauddin Zakariya University, Multan, Pakistan.}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://CRAN.R-project.org/package=DEEVD}
}

}
\author{
Javaria Ahmad Khan, Atif Akbar.
}
